Support login of different users

+++ main.py
@@ -1,18 +1,29 @@
 import datetime
 import sqlite3
 from math import ceil
+from urllib.parse import urlparse, urljoin
 from io import BytesIO
-from flask import Flask, render_template, request, g, send_file, redirect, url_for, make_response, jsonify
+from flask import Flask, render_template, request, g, send_file, redirect, url_for, make_response, jsonify, abort
 from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager, login_user, logout_user, login_required, current_user
 
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:////tmp/test.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.secret_key = b'E\xbdv\xf0\xbd7\x0b\xf9\xce.\x94\xcerx5\xcd'
 db = SQLAlchemy(app)
+login_manager = LoginManager()
+login_manager.init_app(app)
+login_manager.login_view = "login"
 
 
 from models import User, BloodPressure
+
+
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(user_id)
 
 
 @app.route('/')
@@ -21,8 +32,43 @@ def index():
     return redirect(url_for('blood_pressure'))
 
 
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    if request.method == 'POST':
+        try:
+            username = str(request.form['username'])
+            password = str(request.form['password'])
+
+            user = User.query.filter_by(username=username).first()
+            if not user:
+                raise ValueError('Unknown user', username)
+            if not user.password == password:
+                raise ValueError('Invalid password', password)
+
+            login_user(user)
+
+            next = request.args.get('next')
+            if not is_safe_url(next):
+                return abort(400)
+
+            return redirect(next or url_for('index'))
+        except (KeyError, ValueError) as e:
+            app.logger.error('Invalid login: %s', e)
+            return render_template('login.html', message=f'Invalid username or password')
+    else:
+        return render_template('login.html')
+
+
+@app.route("/logout")
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for('login'))
+
+
 @app.route('/blood_pressure', methods=['GET', 'POST'], defaults={'key': None})
 @app.route('/blood_pressure/<int:key>', methods=['DELETE'])
+@login_required
 def blood_pressure(key):
     ROWS_PER_PAGE = 10;
 
@@ -45,7 +91,7 @@ def blood_pressure(key):
             diastolic = int(request.form['diastolic'])
 
             measurement = BloodPressure(
-                user=User.query.first(),  # TODO
+                user=current_user,
                 date=datetime.datetime(date.year, date.month, date.day, hour, minute),
                 systolic=systolic,
                 diastolic=diastolic,
@@ -66,6 +112,9 @@ def blood_pressure(key):
     elif request.method == 'DELETE':
         measurement = BloodPressure.query.get(key)
         if measurement:
+            if measurement.user != current_user:
+                return make_response("You are not allowed to delete this entry", 403)
+
             app.logger.info('Deleting BloodPressure %s', measurement)
             db.session.delete(measurement)
             db.session.commit()
@@ -81,8 +130,7 @@ def blood_pressure(key):
             except:
                 pass
 
-        user = User.query.first()  # TODO
-        rows = BloodPressure.query.with_parent(user)[0:ROWS_PER_PAGE]
+        rows = BloodPressure.query.with_parent(current_user)[0:ROWS_PER_PAGE]
         return render_template(
             'blood_pressure.html',
             header=('Date', 'SYS', 'DIA'),
@@ -95,10 +143,11 @@ def blood_pressure(key):
 
 
 @app.route('/blood_pressure/csv')
+@login_required
 def blood_pressure_csv():
     rows = [
         (row.date.isoformat(), row.systolic, row.diastolic)
-        for row in BloodPressure.query.all()
+        for row in BloodPressure.query.filter_by(user=current_user)
     ]
     return send_csv('blood_pressure', ['date', 'systolic', 'diastolic'], rows)
 
@@ -113,6 +162,12 @@ def send_csv(name, headers, rows):
         csv.write(line.encode('utf-8'))
     csv.seek(0)
     return send_file(csv, mimetype='text/csv', as_attachment=True, attachment_filename=name + '.csv')
+
+
+def is_safe_url(target):
+    ref_url = urlparse(request.host_url)
+    test_url = urlparse(urljoin(request.host_url, target))
+    return test_url.scheme in ('http', 'https') and ref_url.netloc == test_url.netloc
 
 
 if __name__ == '__main__':
